Add parsing of image references
