+++ test/from-markdown/image-def-notfound/input.md
@@ -1,0 +1,1 @@
+This is an image ![1]
