+++ test/from-markdown/image-def-notfound/output.yaml
@@ -1,0 +1,6 @@
+nodes:
+  - kind: block
+    type: paragraph
+    nodes:
+      - kind: text
+        text: "This is an image ![1]"
