+++ test/from-markdown/image-def/input.md
@@ -1,0 +1,3 @@
+This is an image ![a]
+
+[a]: https://www.google.fr/logo.png
