+++ test/from-markdown/image-def/output.yaml
@@ -1,0 +1,13 @@
+nodes:
+  - kind: block
+    type: paragraph
+    nodes:
+      - kind: text
+        text: "This is an image "
+      - kind: inline
+        type: image
+        isVoid: true
+        data:
+          src: https://www.google.fr/logo.png
+      - kind: text
+        text: ""
